Refactor firebase and store data using context Relates #17
