+++ src/App.jsx
@@ -1,18 +1,29 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-
+import { useEffect, useState } from 'react';
 import * as pages from './pages';
 
+import personalData from './context';
+import { getDetails } from './api';
+
 import 'antd/dist/antd.css';
-// import './App.less';
 
-const App = () => (
+const App = () => {
+  const [info, setInfo] = useState({});
 
-  <Router>
-    <Routes>
-      <Route path="/" element={<pages.Main />} />
-    </Routes>
-  </Router>
+  useEffect(async () => {
+    const data = await getDetails();
+    setInfo(data);
+  }, []);
 
-);
+  return (
+    <personalData.Provider value={info}>
+      <Router>
+        <Routes>
+          <Route path="/" element={<pages.Main />} />
+        </Routes>
+      </Router>
+    </personalData.Provider>
+  );
+};
 
 export default App;
