+++ src/context/index.js
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+const personalData = createContext();
+
+export default personalData;
